LayoutTests/imported/w3c/web-platform-tests/trusted-types/trusted-types-reporting-for-Window-setTimeout-setInterval-expected.txt

CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: Refused to connect to ws://common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.

PASS No violation reported for setTimeout with TrustedScript.
PASS No violation reported for setInterval with TrustedScript.
PASS Violation report for setTimeout with plain string.
PASS Violation report for setInterval with plain string.
